contracts,metrics,sqlite: track account metrics
